Refactors material into AbstractRdfSourceOperationBuilder
