+++ fcrepo-kernel-impl/src/main/java/org/fcrepo/kernel/impl/operations/CreateRdfSourceOperationBuilder.java
@@ -17,36 +17,14 @@
  */
 package org.fcrepo.kernel.impl.operations;
 
-import static org.apache.jena.riot.RDFLanguages.contentTypeToLang;
-import static org.fcrepo.kernel.api.rdf.DefaultRdfStream.fromModel;
-
-import org.apache.jena.rdf.model.Model;
-import org.apache.jena.rdf.model.ModelFactory;
-import org.apache.jena.rdf.model.ResourceFactory;
-import org.apache.jena.riot.Lang;
-import org.fcrepo.kernel.api.RdfStream;
 import org.fcrepo.kernel.api.operations.RdfSourceOperation;
-import org.fcrepo.kernel.api.operations.RdfSourceOperationBuilder;
-
-import java.io.InputStream;
-
 
 /**
  * Builder for operations to create rdf sources
  *
  * @author bbpennel
  */
-public class CreateRdfSourceOperationBuilder implements RdfSourceOperationBuilder {
-
-    /**
-     * Holds the stream of user's triples.
-     */
-    private RdfStream tripleStream;
-
-    /**
-     * String of the resource ID.
-     */
-    private final String resourceId;
+public class CreateRdfSourceOperationBuilder extends AbstractRdfSourceOperationBuilder {
 
     /**
      * Constructor.
@@ -54,33 +32,12 @@ public class CreateRdfSourceOperationBuilder implements RdfSourceOperationBuilde
      * @param resourceId the internal identifier.
      */
     public CreateRdfSourceOperationBuilder(final String resourceId) {
-        this.resourceId = resourceId;
+        super(resourceId);
     }
 
     @Override
     public RdfSourceOperation build() {
-        return new CreateRdfSourceOperation(this.resourceId, tripleStream);
+        return new CreateRdfSourceOperation(resourceId, tripleStream);
     }
 
-    @Override
-    public RdfSourceOperationBuilder triples(final RdfStream triples) {
-        this.tripleStream = triples;
-        return this;
-    }
-
-    @Override
-    public RdfSourceOperationBuilder triples(final InputStream contentStream, final String mimetype) {
-        final RdfStream stream;
-        if (contentStream != null && mimetype != null) {
-            final Model model = ModelFactory.createDefaultModel();
-            final Lang lang = contentTypeToLang(mimetype);
-            model.read(contentStream, this.resourceId, lang.getName().toUpperCase());
-            stream = fromModel(ResourceFactory.createResource(this.resourceId).asNode(), model);
-        } else {
-            stream = null;
-        }
-
-        tripleStream = stream;
-        return this;
-    }
 }
