+++ fcrepo-kernel-impl/src/main/java/org/fcrepo/kernel/impl/operations/UpdateRdfSourceOperationBuilder.java
@@ -17,19 +17,7 @@
  */
 package org.fcrepo.kernel.impl.operations;
 
-import java.io.InputStream;
-
-import org.apache.jena.rdf.model.Model;
-import org.apache.jena.rdf.model.ModelFactory;
-import org.apache.jena.rdf.model.ResourceFactory;
-import org.apache.jena.riot.Lang;
-import org.fcrepo.kernel.api.RdfStream;
 import org.fcrepo.kernel.api.operations.RdfSourceOperation;
-import org.fcrepo.kernel.api.operations.RdfSourceOperationBuilder;
-
-import static org.apache.jena.riot.RDFLanguages.contentTypeToLang;
-import static org.fcrepo.kernel.api.rdf.DefaultRdfStream.fromModel;
-
 
 /**
  * Builder for operations to update rdf sources
@@ -39,17 +27,7 @@ import static org.fcrepo.kernel.api.rdf.DefaultRdfStream.fromModel;
  * @since 11/2019
  */
 
-public class UpdateRdfSourceOperationBuilder implements RdfSourceOperationBuilder {
-
-    /**
-     * Holds the stream of user's triples.
-     */
-    private RdfStream tripleStream;
-
-    /**
-     * String of the resource ID.
-     */
-    private final String resourceId;
+public class UpdateRdfSourceOperationBuilder extends AbstractRdfSourceOperationBuilder {
 
     /**
      * Constructor.
@@ -57,34 +35,12 @@ public class UpdateRdfSourceOperationBuilder implements RdfSourceOperationBuilde
      * @param resourceId the internal identifier.
      */
     public UpdateRdfSourceOperationBuilder(final String resourceId) {
-        this.resourceId = resourceId;
+        super(resourceId);
     }
 
     @Override
     public RdfSourceOperation build() {
         // TODO Perform triples validation, relaxed SMTs, etc here?
-        return new UpdateRdfSourceOperation(this.resourceId, this.tripleStream);
-    }
-
-    @Override
-    public RdfSourceOperationBuilder triples(final RdfStream triples) {
-        tripleStream = triples;
-        return this;
-    }
-
-    @Override
-    public RdfSourceOperationBuilder triples(final InputStream contentStream, final String mimetype) {
-        final RdfStream stream;
-        if (contentStream != null && mimetype != null) {
-            final Model model = ModelFactory.createDefaultModel();
-            final Lang lang = contentTypeToLang(mimetype);
-            model.read(contentStream, this.resourceId, lang.getName().toUpperCase());
-            stream = fromModel(ResourceFactory.createResource(this.resourceId).asNode(), model);
-        } else {
-            stream = null;
-        }
-
-        tripleStream = stream;
-        return this;
+        return new UpdateRdfSourceOperation(resourceId, tripleStream);
     }
 }
